Rename models directory and update imports
