+++ lib/models/installment_plan.dart
