+++ lib/models/product.dart
